tracker stats fixes
git-svn-id: http://svn.vuze.com/public/client@3704 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/tracker/host/TRHostTorrent.java
@@ -77,13 +77,13 @@ TRHostTorrent
 	public TRHostPeer[]
 	getPeers();
 	
-	public int
+	public long
 	getAnnounceCount();
 	
-	public int
+	public long
 	getScrapeCount();
 	
-	public int
+	public long
 	getCompletedCount();
 	
 	public long
