+++ trunk/azureus2/src/org/gudy/azureus2/core3/tracker/host/impl/TRHostTorrentPublishImpl.java
@@ -127,18 +127,18 @@ TRHostTorrentPublishImpl
 		return( peers );
 	}	
 
-	public int
+	public long
 	getAnnounceCount()
 	{
 		return( 0 );
 	}
 	
-	public int
+	public long
 	getScrapeCount()
 	{
 		return( 0 );
 	}
-	public int
+	public long
 	getCompletedCount()
 	{
 		return( 0 );
